requirements.txt
tkinter
pillow
transformers
huggingface_hub
torch
pyperclip
tkinterdnd2